+++ examples/intel/ip/ip.py
@@ -2,19 +2,19 @@ import os
 
 import pangea.exceptions as pe
 from pangea.config import PangeaConfig
-from pangea.services import DomainIntel
+from pangea.services import IpIntel
 
-token = os.getenv("PANGEA_DOMAIN_INTEL_TOKEN")
+token = os.getenv("PANGEA_IP_INTEL_TOKEN")
 domain = os.getenv("PANGEA_DOMAIN")
 config = PangeaConfig(domain=domain)
-intel = DomainIntel(token, config=config)
+intel = IpIntel(token, config=config)
 
 
 def main():
-    print(f"Checking domain...")
+    print(f"Checking IP...")
 
     try:
-        response = intel.lookup(domain="737updatesboeing.com", provider="domaintools", verbose=True, raw=True)
+        response = intel.reputation(ip="93.231.182.110", provider="crowdstrike", verbose=True, raw=True)
         print(f"Response: {response.result}")
     except pe.PangeaAPIException as e:
         print(f"Request Error: {e.response.summary}")