+++ packages/pangea-sdk/pangea/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.0"
+__version__ = "1.1.1"
 
 from pangea.config import PangeaConfig
 from pangea.request import PangeaRequest