+++ packages/pangea-sdk/pangea/deep_verify.py
@@ -11,7 +11,8 @@ from itertools import groupby
 
 import pangea.services.audit.util as audit_util
 from pangea.services import Audit
-from pangea.tools import Event, SequenceFollower, exit_with_error, file_events, init_audit, print_progress_bar
+from pangea.services.audit.models import EventEnvelope
+from pangea.tools_util import Event, SequenceFollower, exit_with_error, file_events, init_audit, print_progress_bar
 
 
 class Errors(t.TypedDict):
@@ -77,7 +78,7 @@ def get_path_size(tree_size: int, leaf_index: int) -> int:
 def _tree_size_left(tree_size: int) -> int:
     """if the tree has size tree_size, return the size of the left child"""
     if tree_size <= 1:
-        return 0
+        return tree_size
     return 2 ** (math.ceil(math.log2(tree_size)) - 1)
 
 
@@ -102,10 +103,8 @@ def verify_hash(data: dict, data_hash: str) -> bool:
     """Verify the hash of an event"""
     succeeded = False
     try:
-        data_canon = audit_util.canonicalize_json(data)
-        computed_hash_dec = audit_util.hash_bytes(data_canon)
-        data_hash_dec = audit_util.decode_hash(data_hash)
-        if computed_hash_dec != data_hash_dec:
+        if not audit_util.verify_envelope_hash(EventEnvelope(**data), data_hash):
+            print("Hash failed: ", data)
             raise ValueError("Hash does not match")
         succeeded = True
     except Exception:
@@ -142,7 +141,7 @@ def print_error(msg: str, level: str = "error"):
         dot = "🟡"
     else:
         dot = "🔴"
-    print(f"{dot} {msg:200s}", end="\r")
+    print(f"{dot} {msg:200s}")
 
 
 def deep_verify(audit: Audit, file: io.TextIOWrapper) -> Errors:
@@ -178,7 +177,7 @@ def deep_verify(audit: Audit, file: io.TextIOWrapper) -> Errors:
         cold_indexes.add(leaf_index)
 
         cold_path_size: t.Optional[int] = None
-        hot_out_of_order: set[int] = set()
+        hot_indexes: set[int] = set()
         for i, event in enumerate(events_by_idx):
             cnt += 1
             tree_size = get_tree_size(event)
@@ -204,21 +203,24 @@ def deep_verify(audit: Audit, file: io.TextIOWrapper) -> Errors:
             else:
                 hot_path = path[:-cold_path_size]
                 cold_path = path[-cold_path_size:]
-            
+
             cold_idx = path2index(tree_size, cold_path)
             if cold_idx != leaf_index:
                 errors["wrong_buffer"] += 1
 
             hot_idx = path2index(len(events_by_idx), hot_path)
-            if hot_idx in hot_out_of_order:
-                hot_out_of_order.remove(hot_idx)
-            elif hot_idx != i:
-                hot_out_of_order.add(hot_idx)
+            hot_indexes.add(hot_idx)
 
-        if hot_out_of_order:
-            errors["missing"] += len(hot_out_of_order)
+        hot_indexes_diff = set(range(len(events_by_idx))) - hot_indexes
+        if len(hot_indexes_diff) > 0:
+            errors["missing"] += len(hot_indexes_diff)
+            print(f"missing hot indexes: {hot_indexes_diff}")
+            print(f"hot_indexes: {hot_indexes} ")
+            print(f"events:")
+            for e in events_by_idx:
+                print(e)
             print_error(
-                f"Lines {buffer_lines[0]}-{buffer_lines[1]} ({buffer_lines[1]-buffer_lines[0]}), Buffer #{cold_idx}: {len(hot_out_of_order)} event(s) missing"
+                f"Lines {buffer_lines[0]}-{buffer_lines[1]} ({buffer_lines[1]-buffer_lines[0]}), Buffer #{cold_idx}: {len(hot_indexes_diff)} event(s) missing"
             )
 
     cold_holes = cold_indexes.holes()