+++ CHANGELOG.md
@@ -7,6 +7,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Intel Domain and URL add reputation endpoint that will replace lookup endpoint
+- Intel File add hashReputation() method. Lookup is deprecated deprecated.
+- Intel File add filepathReputation() method. lookupFilepath is deprecated.
+
+
+## [1.1.1] - 2023-01-25
+
+### Changed
+
+- Intel IP add reputation endpoint that will replace lookup endpoint
+- Update User-Agent format
+
+
+## [1.1.0] - 2023-01-05
+
+### Added
+
+- Intel add IP and URL services with lookup endpoint
+
+
+## [1.0.2] - 2022-12-23
+
+### Fixed
+
+- Multiples bugs in audit tools used to dump and verify events
+
+
+## [1.0.1] - 2022-12-19
+
 ### Added
 
 - This CHANGELOG
@@ -41,5 +71,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redact client
 
 
-[unreleased]: https://github.com/pangeacyber/pangea-python/compare/v1.0.0...main
+[unreleased]: https://github.com/pangeacyber/pangea-python/compare/v1.1.1...main
+[1.1.1]: https://github.com/pangeacyber/pangea-python/compare/v1.1.0...v1.1.1
+[1.1.0]: https://github.com/pangeacyber/pangea-python/compare/v1.0.2...v1.1.0
+[1.0.2]: https://github.com/pangeacyber/pangea-python/compare/v1.0.1...v1.0.2
+[1.0.1]: https://github.com/pangeacyber/pangea-python/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/pangeacyber/pangea-python/releases/tag/v1.0.0